AUTH feat(user-store): интеграция zustand user store и рефакторинг главной страницы
- Добавлен user store на zustand (src/store/user.ts) с типами и экшенами setUser/resetUser
- Методы Auth.login/Auth.register теперь типизированы и возвращают объект { user }
- AuthModal теперь сохраняет пользователя в user store после успешного входа
- Вынесена логика главной страницы в компонент MainPage (src/screens/MainPage/MainPage.tsx)
- Главная страница теперь использует user store и отображает email авторизованного пользователя
- src/app/page.tsx теперь просто рендерит <MainPage /> для чистоты архитектуры
- Улучшена модульность: добавлены index.ts для экспорта MainPage

Изменения повышают масштабируемость, типобезопасность и соответствуют архитектурным стандартам проекта.
